+++ veecle-os-runtime/tests/memory_pool.rs
@@ -14,9 +14,12 @@ fn memory_pool() {
 
     static POOL: MemoryPool<u8, 5> = MemoryPool::new();
 
-    #[derive(Debug, Storable)]
-    #[storable(data_type = Chunk<'static, u8>)]
+    #[derive(Debug)]
     pub struct Data;
+
+    impl Storable for Data {
+        type DataType = Chunk<'static, u8>;
+    }
 
     #[veecle_os_runtime::actor]
     async fn exclusive_read_actor(mut reader: ExclusiveReader<'_, Data>) -> Infallible {